+++ src/tfjs/load.ts
@@ -68,7 +68,7 @@ export async function loadModel(modelPath: string | undefined): Promise<GraphMod
     model.loadSync(artifacts); // load weights
     // @ts-ignore private property
     modelStats[shortModelName].sizeLoadedWeights = model.artifacts?.weightData?.byteLength || 0;
-    if (options.verbose) log('load model:', model['modelUrl'], { bytes: modelStats[shortModelName].sizeLoadedWeights }, options);
+    if (options.verbose) log('load:', { model: shortModelName, url: model['modelUrl'], bytes: modelStats[shortModelName].sizeLoadedWeights });
     loaded = true;
   } catch (err) {
     log('error loading model:', modelUrl, err);